scripts/requirements.txt
supabase==2.15.1
python-dotenv==1.1.0
